Make Scroll widget require child to fill bounds
Currently the scroll widget gives its child no minimum size during layout, which means the content of the scroll view can be smaller than the scroll view. This has (I think) unintended consequences, and can be difficult to reason about. By forcing the content to have the same minimum size as the scroll view, we force the user to explicitly align their content or otherwise ensure it fills the available space.